+++ internal/server/event_handlers_test.go
@@ -2,7 +2,10 @@ package server
 
 import (
 	"errors"
+	"strings"
 	"testing"
+
+	"github.com/fluxcd/pkg/runtime/logger"
 )
 
 func TestRedactTokenFromError(t *testing.T) {
@@ -42,17 +45,31 @@ func TestRedactTokenFromError(t *testing.T) {
 			originalErrStr: `Cannot post to github with token metoo8h0387hdyehbwwa45\\n`,
 			expectedErrStr: `Cannot post to github with token metoo*****\\n`,
 		},
+		{
+			name:           "extra text in front token",
+			token:          "8h0387hdyehbwwa45踙",
+			originalErrStr: `Cannot post to github with token metoo8h0387hdyehbwwa45踙\\n`,
+			expectedErrStr: `Cannot post to github with token metoo*****\\n`,
+		},
+		{
+			name:           "return error on invalid UTF-8 string",
+			token:          "\x18\xd0\xfa\xab\xb2\x93\xbb;\xc0l\xf4\xdc",
+			originalErrStr: `Cannot post to github with token \x18\xd0\xfa\xab\xb2\x93\xbb;\xc0l\xf4\xdc\\n`,
+			expectedErrStr: `error redacting token from error message`,
+		},
 	}
 
 	for _, tt := range tests {
-		err := redactTokenFromError(errors.New(tt.originalErrStr), tt.token)
+		log := logger.NewLogger(logger.Options{})
+		err := redactTokenFromError(errors.New(tt.originalErrStr), tt.token, log)
 		if err == nil {
 			t.Fatalf("error shouldn't be nil")
 		}
 
-		if err.Error() != tt.expectedErrStr {
+		if !strings.Contains(err.Error(), tt.expectedErrStr) {
 			t.Errorf("expected error string '%s' but got '%s'",
 				tt.expectedErrStr, err)
 		}
 	}
+
 }